upNext display functional, lag issues fixed,
